chore(graphics): consolidate post render pipeline code
